# server/docs/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "SCIM API documentation for Gluu. For more details please visit https://tools.ietf.org/html/draft-ietf-scim-api-19",
    "version" : "v2",
    "title" : "Gluu SCIM 2.0 Apis Documentation",
    "termsOfService" : "https://github.com/GluuFederation/",
    "contact" : {
      "name" : "Rahat Ali",
      "url" : "http://gluu.org",
      "email" : "rahat@gluu.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host" : "gluu.org",
  "basePath" : "/api",
...
  "schemes" : [ "http", "https" ],
  "paths" : {
...
    "/scim/v2/Groups" : {
      "get" : {
        "tags" : [ "v2Groups" ],
        "summary" : "List Users",
        "description" : "Returns a list of Groups (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.4)",
        "operationId" : "listGroups",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "filter",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortBy",
          "in" : "query",
          "required" : false,
          "type" : "string"
        }, {
          "name" : "sortOrder",
          "in" : "query",
          "required" : false,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/ListResponse"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "post" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Create Group",
        "description" : "Create Group (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "createGroup",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Group",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Group"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    },
    "/scim/v2/Groups/{id}" : {
      "get" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Find Group by id",
        "description" : "Returns a Group on the basis of provided id as path param (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.4.2.1)",
        "operationId" : "getGroupById",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "put" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Update Group",
        "description" : "Update Group (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "updateGroup",
        "consumes" : [ "application/json", "application/xml" ],
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "in" : "body",
          "name" : "body",
          "description" : "Group",
          "required" : false,
          "schema" : {
            "$ref" : "#/definitions/Group"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "$ref" : "#/definitions/Group"
            }
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      },
      "delete" : {
        "tags" : [ "v2Groups" ],
        "summary" : "Delete Group",
        "description" : "Delete Group (https://tools.ietf.org/html/draft-ietf-scim-api-19#section-3.3)",
        "operationId" : "deleteGroup",
        "produces" : [ "application/json", "application/xml" ],
        "parameters" : [ {
          "name" : "Authorization",
          "in" : "header",
          "required" : false,
          "type" : "string",
          "default" : ""
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "default" : {
            "description" : "successful operation"
          }
        },
        "security" : [ {
          "Authorization" : [ ]
        } ]
      }
    },
...
  },
  "definitions" : {
...
    "Group" : {
      "required" : [ "schemas" ],
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "externalId" : {
          "type" : "string"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        },
        "schemas" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "displayName" : {
          "type" : "string"
        },
        "members" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "$ref" : "#/definitions/MemberRef"
          }
        }
      }
    },
...
    "ListResponse" : {
      "properties" : {
        "totalResults" : {
          "type" : "integer",
          "format" : "int32"
        },
        "startIndex" : {
          "type" : "integer",
          "format" : "int32"
        },
        "itemsPerPage" : {
          "type" : "integer",
          "format" : "int32"
        },
        "schemas" : {
          "type" : "array",
          "items" : {
            "type" : "string"
          }
        },
        "resources" : {
          "type" : "array",
          "items" : {
            "$ref" : "#/definitions/Resource"
          }
        }
      },
      "xml" : {
        "name" : "Resources"
      }
    },
    "MemberRef" : {
      "properties" : {
        "operation" : {
          "type" : "string"
        },
        "value" : {
          "type" : "string"
        },
        "display" : {
          "type" : "string"
        },
        "type" : {
          "$ref" : "#/definitions/Type"
        },
        "$ref" : {
          "type" : "string"
        }
      }
    },
    "Meta" : {
      "properties" : {
        "created" : {
          "type" : "string",
          "format" : "date-time"
        },
        "lastModified" : {
          "type" : "string",
          "format" : "date-time"
        },
        "location" : {
          "type" : "string"
        },
        "version" : {
          "type" : "string"
        },
        "attributes" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        },
        "resourceType" : {
          "type" : "string"
        }
      }
    },
...
    "Resource" : {
      "required" : [ "schemas" ],
      "properties" : {
        "id" : {
          "type" : "string"
        },
        "externalId" : {
          "type" : "string"
        },
        "meta" : {
          "$ref" : "#/definitions/Meta"
        },
        "schemas" : {
          "type" : "array",
          "uniqueItems" : true,
          "items" : {
            "type" : "string"
          }
        }
      }
    },
...
    "Type" : {
      "properties" : {
        "value" : {
          "type" : "string"
        }
      }
    },
...
  }
}